+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 We follow [Semantic Versions](https://semver.org/).
 
+## Unreleased
+
+- Add python invocations
+- Add django invocations
+- Add fastapi invocations
+- Add celery invocations
+- Add open-api invocations
+
 ## 0.1.0
 
 - Beta release